# docs/spec/openapi.yaml
openapi: 3.1.0
...
info:
  title: show-gi API
  version: '1.0.0'
  summary: 개입 타이밍을 스스로 정하는 쇼기 학습 상대의 HTTP API
  description: |
    ## 두 가지 규약이 응답 전체를 지배한다

    **실패는 언제나 `{error, message}` 다.** `error` 는 기계용 영어 코드,
    `message` 는 **화면에 그대로 나가는 일본어**다. 클라이언트가 코드로 문장을
    짓지 않는다 — 짓기 시작하면 어휘가 두 벌이 되고 한쪽만 고쳐진 채로 남는다.

    **「없는 기능」과 「지금 못 쓰는 기능」과 「로그인이 필요한 기능」이 갈린다.**

    | 상태 | 응답 |
    | --- | --- |
    | 경로가 아예 없다 | `404` (그 기능이 이 배포에 없다) |
    | DB 가 없다 | `503 store_unavailable` |
    | 엔진이 없다 | `503 engine_unavailable` |
    | 로그인 안 했다 | `401 unauthorized` / `login_required` |
    | 남의 것 · 없는 것 | **같은** `404 not_found` |

    마지막 줄이 의도다. 갈라 주면 id 를 하나씩 올려 보며 남의 기록이 있는지 셀 수 있다.

    ## 관점(perspective)이 칸마다 다르다

    cp 는 부호가 뒤집히는 값이라 어느 쪽 관점인지가 명세의 일부다.

    - DB 저장(`game_moves.eval_cp` · `edges.eval_by_depth`)은 **先手 관점**
    - API 응답의 `evalCp` 는 **플레이어 관점** — 서버가 뒤집어서 준다
    - `/api/explore` 의 `baselineCp` 만 **언제나 下手 관점**이다 (검토에는 플레이어가 없다)

  license:
    name: MIT
    url: https://github.com/jovid18/show-gi/blob/main/LICENSE
servers:
  - url: https://show-gi.com
    description: 프로덕션
  - url: http://localhost:8080
    description: 로컬 (api 컨테이너 직결)
...
security:
  - sessionCookie: []
paths:
...
  /api/auth/google/start:
    get:
      tags: [auth]
      summary: Google 로 보낸다
      description: |
        `state` 를 쿠키에 심고 Google 로 302. **로그인이 꺼져 있으면 경로 자체가 없다(404).**
      security: []
      responses:
        '302':
          description: Google 인증 화면으로
          headers:
            Location: { schema: { type: string, format: uri } }
        '500':
          $ref: '#/components/responses/Error'
...
  /api/me/profile:
    get:
      tags: [review]
      summary: 마이페이지 — 段級·전적·崩れやすいところ
      description: |
        **로그인 안 했으면 401.** 익명 판은 서로 구별할 수단이 없어 「이 사람의 전적」에
        답할 수가 없다.

        집계 셋(`record` · `weaknesses` · `styles`)이 **대인전을 뺀다**(`match_id IS NULL`).
        상대와 둔 판이 그 사람의 「崩れやすいところ」를 흔들지 않는다.
      responses:
        '200':
          description: 그 사람의 段級·전적·崩れやすいところ
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/kifu/parse:
    post:
      tags: [review]
      summary: 밖에서 둔 기보를 읽기만 한다
      description: |
        **판도 안 만들고 엔진도 안 쓴다.** 취해 오기와 갈라 둔 이유가 그것이다 — 잘못
        읽은 기보 하나가 엔진 몇 분이고, 그 사이에 사람이 手数와 앞뒤 수를 눈으로 확인한다.

        **읽는 순서가 규약이다**: USI → CSA → KIF → KI2 → 평문. 하나라도 성공하면
        **OpenAI 를 안 부른다.** 전부 실패했을 때만 그 계층이 서고, 거기서 낸 표기도
        룰 엔진의 전수 검증을 지나야 수가 된다 (journal §126).

        **로그인 안 했으면 401.**
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KifuRequest' }
      responses:
        '200':
          description: 읽은 결과. 판은 아직 없다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/KifuPreview' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/kifu/import:
    post:
      tags: [review]
      summary: 읽은 기보를 판으로 남기고 분석 줄에 세운다
      description: |
        **원문을 다시 보낸다.** 서버에 중간 상태가 없다 — 파싱이 결정적이라 같은 원문이
        같은 결과를 준다.

        `games.result` 는 **주인 관점**이다. 기보의 先手/後手 승패를 `myColor` 로 뒤집고,
        기보가 결과를 안 말하면 `result` 를 요구한다(`400 result`).

        **벽 다섯**: 원문 64KB · 512手 · 사람당 하루 10판(`429 quota`) · 사람당 정규화
        20회/시간 · 분석기가 없으면 `503`. 하루 몫의 벽이 읽기보다 먼저 선다.

        **하루 몫이 정규화를 안 막는다** — 그쪽은 판이 만들어지는 것을 세는데 옮겨
        적는 일은 그 전에 일어난다. 그래서 시간당 벽이 따로 있다.

        판이 남은 뒤에도 평가치는 비어 있다. 채우는 것은 분석 워커이고, 그동안 되짚기가
        `analyzing: true` 를 본다.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/KifuImportRequest' }
      responses:
        '200':
          description: 취해 온 판의 번호. 화면이 그 자리에서 되짚기로 옮겨 간다
          content:
            application/json:
              schema:
                type: object
                required: [gameId]
                properties:
                  gameId: { type: integer, format: int64 }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '429': { $ref: '#/components/responses/Error' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games:
    get:
      tags: [review]
      summary: 최근 대국 목록
      description: |
        **결과가 나온 판만.** 두는 중도 中断도 안 온다.

        익명 판은 쿠키 없이도 읽힌다 — 「자기 판」의 정의가 로그인한 사람에게는
        `user_id`, 익명에게는 `user_id IS NULL` 이다.

        배열이 아니라 객체로 감싼 것은 나중에 커서를 붙일 자리를 비워 둔 것이다.
      security: []
      parameters:
        - name: limit
          in: query
          description: 기본 20, 최대 100 으로 잘린다. `int32` 로 파싱한다
          schema: { type: integer, format: int32, minimum: 1, default: 20 }
      responses:
        '200':
          description: 결과가 나온 자기 판만
          content:
            application/json:
              schema:
                type: object
                required: [games]
                properties:
                  games:
                    type: array
                    items: { $ref: '#/components/schemas/GameSummary' }
        '400':
          description: '`bad_limit` — 0 이하·정수 아님·int32 초과'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games/{id}:
    get:
      tags: [review]
      summary: 한 판 전체 — 手数마다의 국면과 물러진 수
      description: |
        목록과 **같은 조건**이라 끝나지 않은 판은 404 다. 없는 판과 남의 판도 같은 404 다.
      security: []
      parameters: [{ $ref: '#/components/parameters/GameID' }]
      responses:
        '200':
          description: 한 판 전체 — 手数마다의 국면과 물러진 수
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameDetail' }
        '400': { $ref: '#/components/responses/BadID' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games/{id}/summary:
    get:
      tags: [review]
      summary: 그 판의 총평
      description: |
        **기보와 따로 간다** — 화면이 판을 먼저 그린다.

        **대인전에는 총평이 없다(`404 no_summary`).** 세는 것이 개입인데 그것이 없다.
      security: []
      parameters: [{ $ref: '#/components/parameters/GameID' }]
      responses:
        '200':
          description: 그 판의 총평
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GameSummaryPayload' }
        '400': { $ref: '#/components/responses/BadID' }
        '404':
          description: '`not_found` 또는 `no_summary`(대인전)'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games/{id}/quiz:
    get:
      tags: [quiz]
      summary: 그 판에서 뽑은 문항
      description: |
        **정답이 안 실려 온다** — 채점이 서버에 있다.

        `ready: false` 는 **아직 만드는 중**이고 「문항 없음」과 다르다. 생성이 수십 초
        걸려서, 그 사이에 화면이 「問題はありません」을 그리면 거짓이 된다 — 판이 끝나는
        자리에서 문항이 0개여도 행을 남기는 것이 이 값을 위해서다.
      security: []
      parameters: [{ $ref: '#/components/parameters/GameID' }]
      responses:
        '200':
          description: 문항 (정답은 안 실린다)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizPayload' }
        '400': { $ref: '#/components/responses/BadID' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games/{id}/quiz/mate:
    post:
      tags: [quiz]
      summary: 詰み 문항 채점
      description: |
        **내가 낸 수만 보낸다** — 玉方의 응수는 저장된 트리에서 서버가 꺼낸다.
        **엔진을 안 쓴다.**

        `attempt` 를 크게 적어 정답을 살 수 없다. 나가는 것은 `hint` 뿐이고 세 번째
        오답에서만 온다 — 정답 수는 응답에 아예 없다.
      security: []
      parameters: [{ $ref: '#/components/parameters/GameID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moves]
              properties:
                moves:
                  type: array
                  items: { $ref: '#/components/schemas/USIMove' }
                  description: 攻方(나)의 수만. 玉方의 응수는 넣지 않는다
                attempt:
                  type: integer
                  minimum: 1
                  description: 이 문항의 몇 번째 시도인가. **화면이 센다** — 서버에 안 남는다
      responses:
        '200':
          description: 채점 결과
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MateResponse' }
        '400':
          description: '`bad_id` · `bad_request` · `bad_move`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '404':
          description: '`not_found` · `not_ready`(만드는 중) · `no_item`(詰み 문항이 없다)'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games/{id}/quiz/best:
    post:
      tags: [quiz]
      summary: 「최선수는?」 채점
      description: 첫 수만 받는다. **엔진을 안 쓴다.** 정답은 맞혔을 때만 응답에 실린다.
      security: []
      parameters: [{ $ref: '#/components/parameters/GameID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [index, move]
              properties:
                index:
                  type: integer
                  description: |
                    `best[]` 의 몇 번째 문항인가. 手数로 가리키지 않는 이유는 문항이
                    뽑히는 기준이 바뀌어도 이 값의 뜻은 안 바뀌기 때문이다
                move: { $ref: '#/components/schemas/USIMove' }
                attempt: { type: integer, minimum: 1 }
      responses:
        '200':
          description: 채점 결과
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BestResponse' }
        '400':
          description: '`bad_id` · `bad_request` · `bad_move`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '404':
          description: '`not_found` · `not_ready` · `no_item`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/StoreUnavailable' }
  /api/games/{id}/whatif:
    post:
      tags: [engine]
      summary: 가정 수순 한 걸음
      description: |
        **DB 와 엔진 둘 다 필요하다.** 뿌리는 서버가 자기 기록에서 다시 둬서 만든다 —
        판(SFEN)을 안 받는다.

        `POST` 인데 아무것도 안 바꾼다. 본문에 수순 배열을 실어야 해서 `POST` 다.
      security: []
      parameters: [{ $ref: '#/components/parameters/GameID' }]
      requestBody:
        required: true
        description: 본문 상한 8 KiB
        content:
          application/json:
            schema:
              type: object
              required: [ply, moves]
              properties:
                ply:
                  type: integer
                  minimum: 0
                  description: 확정된 몇 手目에서 갈라지나
                moves:
                  type: array
                  maxItems: 60
                  items: { $ref: '#/components/schemas/USIMove' }
      responses:
        '200':
          description: 가정 수순의 지금 자리
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WhatifNode' }
        '400':
          description: '`bad_id` · `bad_request` · `bad_ply` · `bad_move` · `bad_line`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
        '503':
          description: '`engine_unavailable` — 엔진 고장·시한 초과(20s), 또는 배포에 엔진이 없다'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /api/explore:
    post:
      tags: [engine]
      summary: 검토 한 걸음 — 手合割 하나의 0手目에서 시작하는 자유 탐색
      description: |
        **엔진만 필요하고 DB 는 캐시로만 쓴다.** 기록이 없는 배포에서도 그대로 돌고,
        같은 국면을 매번 다시 재는 것만 달라진다.

        **로그인이 필요 없다** (journal §100). 뿌리가 手合割 표라 여는 기록이 아예 없고,
        엔진을 지키는 것은 아래 슬롯 하나다.

        **엔진 슬롯이 1개다.** 풀은 대국이 쓰는 것과 같은 3개인데, 안 묶으면 검토 세 건이
        엔진을 다 잡고 대국의 착수가 그 뒤에 큐에 선다.
      requestBody:
        required: true
        description: 본문 상한 16 KiB
        content:
          application/json:
            schema:
              type: object
              required: [handicap, moves]
              properties:
                handicap:
                  type: string
                  description: '`/api/handicaps` 의 `id`. 빈 문자열이면 平手'
                moves:
                  type: array
                  maxItems: 200
                  description: 0手目부터의 수순. 상한이 가정 수순(60)보다 긴 것은 한 판을 통째로 걸어 볼 수 있어야 해서다
                  items: { $ref: '#/components/schemas/USIMove' }
      responses:
        '200':
          description: 검토의 지금 자리
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExploreNode' }
        '400':
          description: '`bad_request` · `bad_handicap` · `bad_move` · `bad_line`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429':
          description: |
            `busy` — 엔진 슬롯을 3초 기다려도 안 났다. **이 API 에서 429 가 나가는
            자리는 여기 하나다.** 다시 눌러 볼 수 있는 실패라 503 과 갈라 뒀다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '503':
          description: '`engine_unavailable`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
  /api/explore/snapshots:
    get:
      tags: [snapshots]
      summary: 검토에서 저장한 국면 목록
      description: |
        **최근에 저장한 것이 앞이고 개수 상한이 없다.** 목록 질의에 `LIMIT` 을 두면
        지울 수 없는 행이 생긴다.
      responses:
        '200':
          description: 최근에 저장한 것이 앞
          content:
            application/json:
              schema:
                type: object
                required: [snapshots]
                properties:
                  snapshots:
                    type: array
                    items: { $ref: '#/components/schemas/ExploreSnapshot' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/SnapshotStoreUnavailable' }
    post:
      tags: [snapshots]
      summary: 지금 보고 있는 자리를 남긴다
      description: |
        **판(SFEN)을 안 받는다.** 手合割 id 와 수순뿐이고, 저장 전에 룰 엔진으로 한 수씩
        되짚어 본다 — 엔진 탐색이 아니라 합법성 검사라 슬롯을 안 잡는다.

        이름이 비면 서버가 手数로 하나 짓는다. **개수 상한도 이름의 UNIQUE 도 없다** —
        같은 이름 둘은 手数와 저장 시각이 가른다.
      requestBody:
        required: true
        description: 본문 상한 16 KiB
        content:
          application/json:
            schema:
              type: object
              required: [handicap, moves]
              properties:
                name:
                  type: string
                  description: 40 rune 까지. **바이트가 아니라 rune 으로 센다.** 비면 서버가 짓는다
                handicap: { type: string }
                moves:
                  type: array
                  maxItems: 200
                  items: { $ref: '#/components/schemas/USIMove' }
      responses:
        '201':
          description: 남겼다
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExploreSnapshot' }
        '400':
          description: '`bad_request` · `bad_name` · `bad_handicap` · `bad_move` · `bad_line`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/SnapshotStoreUnavailable' }
  /api/explore/snapshots/{id}:
    patch:
      tags: [snapshots]
      summary: 이름만 고친다
      description: |
        **국면은 안 바뀐다.** 수순을 같은 행에 덮어쓰면 옛 이름이 가리키던 자리가 조용히
        다른 국면이 된다.

        없는 것과 남의 것이 **같은 404** 다.
      parameters: [{ $ref: '#/components/parameters/SnapshotID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, description: 1~40 rune }
      responses:
        '200':
          description: 이름을 고쳤다
          content:
            application/json:
              schema:
                type: object
                required: [id, name]
                properties:
                  id: { type: integer, format: int64 }
                  name: { type: string }
        '400':
          description: '`bad_id` · `bad_request` · `bad_name`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/SnapshotStoreUnavailable' }
    delete:
      tags: [snapshots]
      summary: 저장된 국면 하나를 지운다
      parameters: [{ $ref: '#/components/parameters/SnapshotID' }]
      responses:
        '204': { description: 지웠다 }
        '400': { $ref: '#/components/responses/BadID' }
        '401': { $ref: '#/components/responses/LoginRequired' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/SnapshotStoreUnavailable' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: |
        **서명 쿠키 하나.** `HttpOnly` 이고 세션이 표가 아니라 서명된 값이라 서버에
        지울 것이 없다. Bearer 토큰도 API 키도 없다 — 클라이언트가 브라우저 하나뿐이다.
...
  responses:
    Error:
      description: '`internal` — 서버 문제. 로그에 남는다'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    LoginRequired:
      description: '`unauthorized` 또는 `login_required`'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    StoreUnavailable:
      description: |
        `store_unavailable` — 이 배포에 DB 가 없다. **대국은 된다** — 엔진이 죽어
        대국을 못 해도 지난 판은 볼 수 있어야 하고, 그 반대도 같다
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...